fix: Rename `Change package name` patch to `Clone app`
